Разбила редьюсер на несколько
