+++ src/components/room-screen/room-screen.jsx
@@ -139,7 +139,7 @@ RoomScreen.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  offers: state.offers,
+  offers: state.DATA.offers,
 });
 
 export {RoomScreen};
