+++ src/store/actions.js
@@ -3,13 +3,12 @@ export const ActionType = {
   SET_SELECTED_CITY_ACTION: `SET_SELECTED_CITY_ACTION`,
 };
 
-export const ActionCreator = {
-  setOffersAction: (payload) => ({
-    type: ActionType.SET_OFFERS_ACTION,
-    payload,
-  }),
-  setSelectedCityAction: (payload) => ({
-    type: ActionType.SET_SELECTED_CITY_ACTION,
-    payload,
-  }),
-};
+export const setOffersAction = (payload) => ({
+  type: ActionType.SET_OFFERS_ACTION,
+  payload,
+});
+
+export const setSelectedCityAction = (payload) => ({
+  type: ActionType.SET_SELECTED_CITY_ACTION,
+  payload,
+});
