+++ src/store/reducers/process.js
@@ -1,0 +1,20 @@
+import {extend} from "../../utils";
+import {ActionType} from "../actions";
+import cities from "../../mocks/cities";
+
+const initialState = {
+  activeCity: cities[0],
+};
+
+const process = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.SET_SELECTED_CITY_ACTION:
+      return extend(state, {
+        activeCity: action.payload,
+      });
+  }
+
+  return state;
+};
+
+export {process};
